server/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioresponses==0.7.6
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
Authlib==1.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
bitarray==3.6.0
blinker==1.9.0
Bootstrap-Flask==2.3.3
bs4==0.0.2
certifi==2024.8.30
cffi==1.16.0
charset-normalizer==3.4.0
ckzg==2.1.1
click==8.2.1
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
cytoolz==1.0.1
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.4
dnspython==2.6.1
dominate==2.9.1
elevenlabs==1.7.0
email-validator==2.1.0.post1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
executing==2.0.1
filelock==3.13.4
Flask==3.1.1
Flask-CKEditor==0.5.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SocketIO==3.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.9.0
gast==0.6.0
gevent==24.2.1
gevent-websocket==0.10.1
git-filter-repo==2.45.0
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.68.1
gTTS==2.5.3
gunicorn==19.9.0
h11==0.14.0
h5py==3.12.1
hexbytes==1.3.1
hijri-converter==2.3.1
httpcore==1.0.5
httpx==0.27.2
idna==3.10
inference-sdk==0.9.17
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.3
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
pandoc==2.4
parsimonious==0.10.0
parso==0.8.4
passlib==1.7.1
paystackapi==2.1.2
phonenumbers==8.13.35
pillow==10.2.0
platformdirs==4.2.1
plumbum==1.9.0
ply==3.11
pooch==1.8.2
prompt-toolkit==3.0.43
protobuf==5.29.1
psycopg2-binary==2.9.9
pure-eval==0.2.2
py-cpuinfo==9.0.0
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.9.0
pydantic_core==2.23.2
pygame==2.6.0
Pygments==2.18.0
pyotp==2.9.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==3.5.1
python-socketio==3.1.2
pytz==2024.1
pyunormalize==16.0.0
pywin32==308
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.3
rich==13.9.4
rlp==4.1.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
selenium==4.21.0
setuptools==75.6.0
simple-websocket==1.0.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.5.0.post1
SQLAlchemy==2.0.41
sqlparse==0.5.0
stack-data==0.6.3
supervision==0.22.0
sympy==1.13.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
toolz==1.0.0
torch==2.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.2.3
virtualenv==20.26.0
visitor==0.1.3
wcwidth==0.2.13
web3==7.13.0
websockets==13.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
zope.event==5.0
zope.interface==6.4.post2
